some foxes
