+++ dapp/dev.env
@@ -40,3 +40,4 @@ DISABLE_WRAP_BUTTON=false
 PINATA_API_KEY=
 PINATA_SECRET_API_KEY=
 IPFS_DAPP_URL=https://ousd.eth.link/swap.html
+AIRDROP_URL=https://governance.ousd.com/claim
