+++ dapp/src/utils/constants.js
@@ -1,3 +1,5 @@
+import moment from 'moment'
+
 const mintAbsoluteGasLimitBuffer = 20000
 /* All gas limit buffers are expressed in relative percentages. Meaning a 0.2
  * value will set gasLimit to 120% of the original value.
@@ -19,6 +21,14 @@ const DEFAULT_SELECTED_APY = 30
 const transactionHistoryItemsPerPage = 50
 const max_price = 1.2
 
+const burnTimer = () => {
+  const burn = moment('2022-10-10T00:00:00.000Z')
+  const days = burn.diff(moment(), 'days')
+  const seconds = burn.diff(moment(), 'seconds')
+  const burnDays = days === 0 ? 1 : days
+  return { days: burnDays, seconds: seconds }
+}
+
 module.exports = {
   mintAbsoluteGasLimitBuffer,
   mintPercentGasLimitBuffer,
@@ -32,4 +42,5 @@ module.exports = {
   DEFAULT_SELECTED_APY,
   transactionHistoryItemsPerPage,
   max_price,
+  burnTimer,
 }
